(Refactoring) Extract CR struct fields parser from renderer for re-usage
